Add files via upload
This is a project created as part of a web development module for a master's course in software development. A fake commerce website selling digital video games was created. A dataset was used containing prices, descriptions, image urls etc from the steam platform. This project combined front end and back end development and involved the creation of a REST API. The final mark awarded for this project was 77.